requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.7.0
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
redis==5.2.1
httpx==0.28.1
python-jose[cryptography]==3.3.0
cryptography==44.0.0
itsdangerous==2.2.0
python-multipart==0.0.20
pygit2==1.16.0
gitpython==3.1.43
kubernetes==31.0.0
jinja2==3.1.5
pyyaml==6.0.2
python-dotenv==1.0.1
structlog==24.4.0
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
prometheus-client==0.21.1
slowapi==0.1.9
starlette-csrf==3.0.0
limits==3.14.0
aiofiles==24.1.0
tenacity==9.0.0
"fastapi[standard]"